feat(backend): start caching user collections list
