+++ crates/services/user/Cargo.toml
@@ -13,6 +13,7 @@ common-utils = { path = "../../utils/common" }
 database-models = { path = "../../models/database" }
 database-utils = { path = "../../utils/database" }
 dependent-models = { path = "../../models/dependent" }
+dependent-utils = { path = "../../utils/dependent" }
 enum_meta = { workspace = true }
 enum-models = { path = "../../models/enum" }
 fitness-models = { path = "../../models/fitness" }
